+++ cpp/src/arrow/compute/CMakeLists.txt
@@ -82,7 +82,8 @@ add_arrow_compute_test(internals_test
                        exec_test.cc
                        kernel_test.cc
                        light_array_test.cc
-                       registry_test.cc)
+                       registry_test.cc
+                       key_hash_test.cc)
 
 add_arrow_benchmark(function_benchmark PREFIX "arrow-compute")
