+++ cpp/src/arrow/compute/exec/swiss_join.cc
@@ -22,7 +22,7 @@
 #include <mutex>
 #include "arrow/array/util.h"  // MakeArrayFromScalar
 #include "arrow/compute/exec/hash_join.h"
-#include "arrow/compute/exec/key_hash.h"
+#include "arrow/compute/key_hash.h"
 #include "arrow/compute/exec/swiss_join_internal.h"
 #include "arrow/compute/exec/util.h"
 #include "arrow/compute/kernels/row_encoder_internal.h"
