+++ cpp/src/arrow/compute/key_hash.h
