+++ cpp/src/arrow/compute/key_hash_avx2.cc
@@ -17,7 +17,7 @@
 
 #include <immintrin.h>
 
-#include "arrow/compute/exec/key_hash.h"
+#include "arrow/compute/key_hash.h"
 #include "arrow/util/bit_util.h"
 
 namespace arrow {
