+++ cpp/src/arrow/compute/key_hash_test.cc
@@ -21,7 +21,7 @@
 #include <map>
 #include <unordered_set>
 #include "arrow/array/builder_binary.h"
-#include "arrow/compute/exec/key_hash.h"
+#include "arrow/compute/key_hash.h"
 #include "arrow/compute/exec/test_util.h"
 #include "arrow/compute/exec/util.h"
 #include "arrow/util/cpu_info.h"
