+++ cpp/src/arrow/compute/light_array.h
@@ -21,7 +21,8 @@
 
 #include "arrow/array.h"
 #include "arrow/compute/exec.h"
-#include "arrow/compute/exec/util.h"
+// DIPO #include "arrow/compute/exec/util.h"
+#include "arrow/compute/util.h"
 #include "arrow/type.h"
 #include "arrow/util/cpu_info.h"
 #include "arrow/util/logging.h"
